feat(db): сбросить схему на PostgreSQL baseline

+++ src/pybot/db/models/user_module/competence.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from sqlalchemy import Integer, Text, UniqueConstraint
+from sqlalchemy import BigInteger, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from ...base_class import Base
@@ -11,7 +11,7 @@ class Competence(Base):
     __tablename__ = "competencies"
     __table_args__ = (UniqueConstraint("name", name="uq_competencies_name"),)
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     name: Mapped[str] = mapped_column(Text, nullable=False)
     description: Mapped[str | None] = mapped_column(Text)
